Extract scanner I/O helper implementations
