+++ constraints.txt
@@ -1,0 +1,10 @@
+#
+# This file is autogenerated by pip-compile with Python 3.13
+# by the following command:
+#
+#    pip-compile --output-file=constraints.txt --strip-extras requirements.txt
+#
+pymodbus==3.13.0
+    # via -r requirements.txt
+voluptuous==0.16.0
+    # via -r requirements.txt
