+++ custom_components/thessla_green_modbus/mappings/_helpers.py
@@ -1,0 +1,19 @@
+"""Helper function exports for mapping generation."""
+
+from __future__ import annotations
+
+from . import (
+    _get_register_info,
+    _infer_icon,
+    _load_translation_keys,
+    _number_translation_keys,
+    _parse_states,
+)
+
+__all__ = [
+    "_get_register_info",
+    "_infer_icon",
+    "_load_translation_keys",
+    "_number_translation_keys",
+    "_parse_states",
+]
