+++ custom_components/thessla_green_modbus/mappings/_loaders.py
@@ -1,0 +1,17 @@
+"""Loader function exports for mapping generation."""
+
+from __future__ import annotations
+
+from . import (
+    _build_entity_mappings,
+    _extend_entity_mappings_from_registers,
+    _load_discrete_mappings,
+    _load_number_mappings,
+)
+
+__all__ = [
+    "_build_entity_mappings",
+    "_extend_entity_mappings_from_registers",
+    "_load_discrete_mappings",
+    "_load_number_mappings",
+]
