+++ custom_components/thessla_green_modbus/mappings/legacy.py
@@ -1,0 +1,7 @@
+"""Legacy entity-id alias exports."""
+
+from __future__ import annotations
+
+from . import LEGACY_ENTITY_ID_ALIASES, map_legacy_entity_id
+
+__all__ = ["LEGACY_ENTITY_ID_ALIASES", "map_legacy_entity_id"]
