+++ custom_components/thessla_green_modbus/mappings/special_modes.py
@@ -1,0 +1,7 @@
+"""Special mode constants exports."""
+
+from __future__ import annotations
+
+from . import SPECIAL_MODE_ICONS
+
+__all__ = ["SPECIAL_MODE_ICONS"]
